requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.2
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-filter==23.2
django-import-export==3.2.0
django-multiselectfield==0.1.12
django-pandas==0.6.6
django-related-admin==1.0.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
filelock==3.12.1
idna==3.4
MarkupPy==1.14
mysqlclient==2.1.1
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.3
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
tablib==3.5.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.0
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
